backend/requirements.txt
# Python核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
mysqlclient==2.2.0
pymysql==1.1.0
aiomysql==0.2.0
asyncpg==0.29.0

# Redis缓存
redis==5.0.1
hiredis==2.2.3

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
pycryptodome==3.19.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 异步任务
celery==5.3.4

# 数据处理
pandas==2.1.4
numpy==1.24.4

# 科学计算（用于预测算法）
scikit-learn==1.3.2
scipy==1.11.4

# 可视化（如果需要）
matplotlib==3.8.2
seaborn==0.13.0

# 日志和监控
loguru==0.7.2
python-json-logger==2.0.7
prometheus-client==0.19.0

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # 用于测试HTTP请求

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 开发工具
python-dotenv==1.0.0
watchfiles==0.21.0

# 微信支付
wechatpay-python==2.0.0

# 其他工具
python-slugify==8.0.1
Pillow==10.1.0
pytz==2023.3